requirements.txt
py-cord[voice]==2.6.1
google-genai
pydantic-settings
PyNaCl